chore(ton): split adapter into constants, utils, parser modules (#12524)
