+++ packages/chain-adapters/src/ton/TonChainAdapter.ts
@@ -1,5 +1,5 @@
 import type { AssetId, ChainId } from '@shapeshiftoss/caip'
-import { ASSET_REFERENCE, toAssetId, tonAssetId, tonChainId } from '@shapeshiftoss/caip'
+import { ASSET_REFERENCE, tonAssetId, tonChainId } from '@shapeshiftoss/caip'
 import type { HDWallet, TonWallet } from '@shapeshiftoss/hdwallet-core'
 import type { Bip44Params, RootBip44Params } from '@shapeshiftoss/types'
 import { KnownChainIds } from '@shapeshiftoss/types'
@@ -24,359 +24,51 @@ import type {
   Transaction,
   TxHistoryInput,
   TxHistoryResponse,
-  TxTransfer,
   ValidAddressResult,
 } from '../types'
 import { ChainAdapterDisplayName, ValidAddressResultType } from '../types'
 import { toAddressNList, verifyLedgerAppOpen } from '../utils'
+import {
+  PARSE_TX_CACHE_TTL_MS,
+  TON_REQUEST_QUEUE_INTERVAL_MS,
+  TRACE_BATCH_SIZE,
+  TRACE_COMPLETION_LT_SPAN,
+  TRACE_LT_SEARCH_RANGE,
+} from './constants'
+import { buildTonTokens, buildTraceTransfers, ownTraceTxs, parseTonTx } from './parser'
 import type {
-  JettonTransferRecord,
+  ChainAdapterArgs,
+  TonAccountInfo,
   TonApiTxResponse,
+  TonConfigParamResult,
   TonFeeData,
+  TonJettonTransfersResponse,
+  TonJettonWalletsResponse,
+  TonMessagesResponse,
+  TonRpcResponse,
+  TonRunGetMethodResult,
+  TonSendBocResult,
   TonSignTx,
   TonToken,
-  TonTrace,
   TonTracesResponse,
   TonTx,
 } from './types'
+import {
+  addressesMatch,
+  formatTonError,
+  getTraceOwnerTxid,
+  isHexHash,
+  isRetryableError,
+  resolveAddresses,
+} from './utils'
 
 const supportsTon = (wallet: HDWallet): wallet is TonWallet => {
   return '_supportsTon' in wallet && (wallet as TonWallet)._supportsTon === true
 }
 
-export type ChainAdapterArgs = {
-  rpcUrl: string
-}
-
-type TonRpcResponse<T> = {
-  ok: boolean
-  result?: T
-  error?: string
-}
-
-type TonAccountInfo = {
-  balance: string
-  state: 'active' | 'uninitialized' | 'frozen'
-  code?: string
-  data?: string
-}
-
-const PROXY_TON_CONTRACTS = new Set([
-  'EQCM3B12QK1e4yZSf8GtBRT0aLMNyEsBc_DhVfRRtOEffLez',
-  'EQBnGWMCf3-FZZq1W4IWcWiGAc3PHuZ0_H-7sad2oY00o83S',
-])
-
-// Logical time advances ~1e6 per second, and downstream trace legs (dex payouts, excesses) land
-// tens of millions of lts after the initiator - this bounds the search only, results are
-// filtered by trace_id
-const TRACE_LT_SEARCH_RANGE = 1_000_000_000n
-// Legs of a trace land within this window of its initiator (~5 minutes of logical time)
-const TRACE_COMPLETION_LT_SPAN = 300_000_000n
-const TRACE_BATCH_SIZE = 20
-// Toncenter free tier allows ~1 req/sec; 429s are retried with Retry-After
-const TON_REQUEST_QUEUE_INTERVAL_MS = 1_100
-// Shorter than the 5s status poll interval so each poll still observes fresh chain state
-const PARSE_TX_CACHE_TTL_MS = 4_000
 // TTL starts at resolution - pending parses are reused as-is so a queue backlog can't
 // trigger duplicate network runs for the same hash
 type ParseTxCacheEntry = { resolvedAt?: number; promise: Promise<Transaction> }
-const TON_HASH_HEX_LENGTH = 64
-export const isHexHash = (str: string): boolean => {
-  return str.length === TON_HASH_HEX_LENGTH && /^[0-9a-f]+$/i.test(str)
-}
-
-export const addressesMatch = (addr1: string, addr2: string): boolean => {
-  if (!addr1 || !addr2) return false
-  if (addr1 === addr2) return true
-  try {
-    return Address.parse(addr1).equals(Address.parse(addr2))
-  } catch {
-    const normalize = (a: string) => a.replace(/^0:/, '').toLowerCase()
-    return normalize(addr1) === normalize(addr2)
-  }
-}
-
-export const isProxyTon = (jettonMaster: string): boolean => {
-  if (PROXY_TON_CONTRACTS.has(jettonMaster)) return true
-  try {
-    const parsed = Address.parse(jettonMaster)
-    for (const known of PROXY_TON_CONTRACTS) {
-      try {
-        if (parsed.equals(Address.parse(known))) return true
-      } catch {
-        continue
-      }
-    }
-  } catch {}
-  return false
-}
-
-export const resolveAddresses = (
-  tx: TonTx,
-  addressBook: Record<string, { user_friendly: string }>,
-): TonTx => {
-  const resolve = (addr: string | undefined): string | undefined =>
-    addr ? addressBook[addr]?.user_friendly ?? addr : addr
-
-  return {
-    ...tx,
-    in_msg: tx.in_msg
-      ? {
-          ...tx.in_msg,
-          source: resolve(tx.in_msg.source),
-          destination: resolve(tx.in_msg.destination),
-        }
-      : tx.in_msg,
-    out_msgs: tx.out_msgs?.map(msg => ({
-      ...msg,
-      source: resolve(msg.source),
-      destination: resolve(msg.destination),
-    })),
-  }
-}
-
-export const buildJettonTransfers = (
-  jettonTransfers: JettonTransferRecord[],
-  traceId: string,
-  pubkey: string,
-  addressBook: Record<string, { user_friendly: string }>,
-  chainId: ChainId,
-): TxTransfer[] => {
-  const transfers: TxTransfer[] = []
-
-  const matching = jettonTransfers.filter(jt => jt.trace_id === traceId)
-  if (matching.length === 0) return transfers
-
-  const friendly = (addr: string) => addressBook[addr]?.user_friendly ?? addr
-
-  for (const transfer of matching) {
-    if (!transfer.source || !transfer.destination || !transfer.amount || !transfer.jetton_master)
-      continue
-
-    const sourceUserFriendly = friendly(transfer.source)
-    const destUserFriendly = friendly(transfer.destination)
-    const jettonUserFriendly = friendly(transfer.jetton_master)
-
-    if (isProxyTon(jettonUserFriendly)) continue
-
-    const isSend = addressesMatch(sourceUserFriendly, pubkey)
-    const isReceive = addressesMatch(destUserFriendly, pubkey)
-
-    if (!isSend && !isReceive) continue
-
-    const assetId = toAssetId({
-      chainId,
-      assetNamespace: 'jetton',
-      assetReference: jettonUserFriendly,
-    })
-
-    if (isSend) {
-      transfers.push({
-        assetId,
-        from: [sourceUserFriendly],
-        to: [destUserFriendly],
-        type: TransferType.Send,
-        value: transfer.amount,
-      })
-    }
-
-    if (isReceive) {
-      transfers.push({
-        assetId,
-        from: [sourceUserFriendly],
-        to: [destUserFriendly],
-        type: TransferType.Receive,
-        value: transfer.amount,
-      })
-    }
-  }
-
-  return transfers
-}
-
-// Raw message values misstate native swap legs (gas budgets ride the envelope, refunds ride the
-// payout) - swap rows use the proxy TON jetton amount when present, net native flow otherwise
-export const buildTraceTransfers = ({
-  txs,
-  jettonTransfers,
-  traceId,
-  pubkey,
-  addressBook,
-  assetId,
-  chainId,
-}: {
-  txs: TonTx[]
-  jettonTransfers: JettonTransferRecord[]
-  traceId: string
-  pubkey: string
-  addressBook: Record<string, { user_friendly: string }>
-  assetId: AssetId
-  chainId: ChainId
-}): TxTransfer[] => {
-  const jetton = buildJettonTransfers(jettonTransfers, traceId, pubkey, addressBook, chainId)
-
-  const seen = new Set<string>()
-  const native: TxTransfer[] = []
-  for (const tx of txs) {
-    const parsed = parseTonTx(resolveAddresses(tx, addressBook), pubkey, '', assetId, chainId)
-    for (const transfer of parsed.transfers) {
-      // Scoped to the source tx so identical legs from different txs both survive
-      const key = `${tx.hash}-${transfer.assetId}-${transfer.from[0]}-${transfer.to[0]}-${transfer.value}-${transfer.type}`
-      if (seen.has(key)) continue
-      seen.add(key)
-      native.push(transfer)
-    }
-  }
-
-  // Plain native transfers keep their per-leg values
-  if (jetton.length === 0) return native
-
-  const friendly = (addr: string) => addressBook[addr]?.user_friendly ?? addr
-
-  const proxyAmounts: Partial<Record<TransferType, string>> = {}
-  for (const transfer of jettonTransfers) {
-    if (transfer.trace_id !== traceId) continue
-    if (!transfer.source || !transfer.destination || !transfer.amount || !transfer.jetton_master)
-      continue
-    if (!isProxyTon(friendly(transfer.jetton_master))) continue
-
-    // Summed per direction - split routes move the wrapped amount in multiple legs
-    if (addressesMatch(friendly(transfer.source), pubkey)) {
-      proxyAmounts[TransferType.Send] = (
-        BigInt(proxyAmounts[TransferType.Send] ?? '0') + BigInt(transfer.amount)
-      ).toString()
-    }
-    if (addressesMatch(friendly(transfer.destination), pubkey)) {
-      proxyAmounts[TransferType.Receive] = (
-        BigInt(proxyAmounts[TransferType.Receive] ?? '0') + BigInt(transfer.amount)
-      ).toString()
-    }
-  }
-
-  // Net native flow across the trace, gas envelopes and excess refunds included (fees excluded)
-  let net = 0n
-  for (const tx of txs) {
-    if (tx.in_msg?.value && tx.in_msg.source) net += BigInt(tx.in_msg.value)
-    for (const outMsg of tx.out_msgs ?? []) {
-      if (outMsg.value) net -= BigInt(outMsg.value)
-    }
-  }
-
-  const hasJettonSend = jetton.some(t => t.type === TransferType.Send)
-  const hasJettonReceive = jetton.some(t => t.type === TransferType.Receive)
-  const sends = native.filter(t => t.type === TransferType.Send)
-  const receives = native.filter(t => t.type === TransferType.Receive)
-
-  const nativeLegs: TxTransfer[] = []
-
-  if (proxyAmounts[TransferType.Send] && sends.length === 1) {
-    nativeLegs.push({ ...sends[0], value: proxyAmounts[TransferType.Send] })
-  } else if (net < 0n && hasJettonReceive && !hasJettonSend && sends.length > 0) {
-    nativeLegs.push({ ...sends[0], value: (-net).toString() })
-  }
-
-  if (proxyAmounts[TransferType.Receive] && receives.length === 1) {
-    nativeLegs.push({ ...receives[0], value: proxyAmounts[TransferType.Receive] })
-  } else if (net > 0n && hasJettonSend && !hasJettonReceive && receives.length > 0) {
-    nativeLegs.push({ ...receives[0], value: net.toString() })
-  }
-
-  return [...nativeLegs, ...jetton]
-}
-
-export const parseTonTx = (
-  tx: TonTx,
-  pubkey: string,
-  txid: string,
-  assetId: AssetId,
-  chainId: ChainId,
-): Transaction => {
-  const isAborted = tx.description?.aborted ?? false
-  const actionSuccess = tx.description?.action?.success ?? true
-  const status = isAborted || !actionSuccess ? TxStatus.Failed : TxStatus.Confirmed
-
-  const transfers: TxTransfer[] = []
-
-  if (tx.in_msg?.value && tx.in_msg.source && tx.in_msg.destination) {
-    const inMsgDecodedType = tx.in_msg.message_content?.decoded?.['@type']
-    const isExcess = inMsgDecodedType === 'excess'
-    const value = tx.in_msg.value
-    if (BigInt(value) > 0n && !isExcess) {
-      const isReceive = addressesMatch(tx.in_msg.destination, pubkey)
-      const isSend = addressesMatch(tx.in_msg.source, pubkey)
-
-      if (isSend) {
-        transfers.push({
-          assetId,
-          from: [tx.in_msg.source],
-          to: [tx.in_msg.destination],
-          type: TransferType.Send,
-          value,
-        })
-      }
-      if (isReceive) {
-        transfers.push({
-          assetId,
-          from: [tx.in_msg.source],
-          to: [tx.in_msg.destination],
-          type: TransferType.Receive,
-          value,
-        })
-      }
-    }
-  }
-
-  if (tx.out_msgs) {
-    for (const outMsg of tx.out_msgs) {
-      if (outMsg.value && outMsg.source && outMsg.destination) {
-        const decodedType = outMsg.message_content?.decoded?.['@type']
-        const value =
-          decodedType === 'pton_ton_transfer' &&
-          outMsg.message_content?.decoded?.ton_amount?.amount?.value
-            ? outMsg.message_content.decoded.ton_amount.amount.value
-            : outMsg.value
-        if (BigInt(value) > 0n) {
-          const isSend = addressesMatch(outMsg.source, pubkey)
-          const isReceive = addressesMatch(outMsg.destination, pubkey)
-
-          if (isSend) {
-            transfers.push({
-              assetId,
-              from: [outMsg.source],
-              to: [outMsg.destination],
-              type: TransferType.Send,
-              value,
-            })
-          }
-          if (isReceive) {
-            transfers.push({
-              assetId,
-              from: [outMsg.source],
-              to: [outMsg.destination],
-              type: TransferType.Receive,
-              value,
-            })
-          }
-        }
-      }
-    }
-  }
-
-  const isSend = transfers.some(transfer => transfer.type === TransferType.Send)
-
-  return {
-    txid,
-    blockHeight: Number(tx.lt) || 0,
-    blockTime: tx.now || 0,
-    blockHash: undefined,
-    chainId,
-    confirmations: status === TxStatus.Confirmed ? 1 : 0,
-    status,
-    transfers,
-    pubkey,
-    ...(isSend && tx.total_fees && { fee: { assetId, value: tx.total_fees } }),
-  }
-}
 
 export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
   static readonly rootBip44Params: RootBip44Params = {
@@ -446,8 +138,8 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
             const data = (await response.json()) as TonRpcResponse<T>
 
             if (!data.ok && data.error) {
-              lastError = new Error(this.formatTonError(data.error))
-              if (this.isRetryableError(data.error)) {
+              lastError = new Error(formatTonError(data.error))
+              if (isRetryableError(data.error)) {
                 const backoffDelay = 1000 * Math.pow(2, attempt)
                 await new Promise(resolve => setTimeout(resolve, backoffDelay))
                 continue
@@ -476,53 +168,6 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
       },
       { throwOnTimeout: true },
     )
-  }
-
-  private formatTonError(error: string): string {
-    if (error.includes('INVALID_BAG_OF_CELLS')) {
-      return `TON transaction serialization error: ${error}. This may indicate an invalid transaction format.`
-    }
-    if (error.includes('seqno')) {
-      return `TON sequence number error: ${error}. The transaction may be stale or already processed.`
-    }
-    if (error.includes('not enough balance') || error.includes('insufficient')) {
-      return `TON insufficient balance: ${error}`
-    }
-    return `TON RPC error: ${error}`
-  }
-
-  private isRetryableError(error: string): boolean {
-    const lowerError = error.toLowerCase()
-
-    const nonRetryablePatterns = [
-      'insufficient',
-      'not enough balance',
-      'invalid',
-      'malformed',
-      'unauthorized',
-      'forbidden',
-      'not found',
-      'bad request',
-      'seqno',
-    ]
-    if (nonRetryablePatterns.some(pattern => lowerError.includes(pattern))) {
-      return false
-    }
-
-    const retryablePatterns = [
-      'timeout',
-      'etimedout',
-      'econnreset',
-      'econnrefused',
-      'network',
-      'temporarily unavailable',
-      'rate limit',
-      '429',
-      '500',
-      '502',
-      '503',
-    ]
-    return retryablePatterns.some(pattern => lowerError.includes(pattern))
   }
 
   private httpApiRequest<T>(endpoint: string): Promise<T> {
@@ -636,62 +281,11 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
       }
 
       try {
-        const jettonsResponse = await this.httpApiRequest<{
-          jetton_wallets?: {
-            address: string
-            balance: string
-            jetton: string
-          }[]
-          address_book?: Record<
-            string,
-            {
-              user_friendly: string
-            }
-          >
-          metadata?: Record<
-            string,
-            {
-              token_info?: {
-                name?: string
-                symbol?: string
-                extra?: {
-                  decimals?: string
-                }
-              }[]
-            }
-          >
-        }>(`/api/v3/jetton/wallets?owner_address=${encodeURIComponent(pubkey)}`)
+        const jettonsResponse = await this.httpApiRequest<TonJettonWalletsResponse>(
+          `/api/v3/jetton/wallets?owner_address=${encodeURIComponent(pubkey)}`,
+        )
 
-        if (jettonsResponse.jetton_wallets) {
-          const addressBook = jettonsResponse.address_book ?? {}
-          const metadata = jettonsResponse.metadata ?? {}
-
-          tokens = jettonsResponse.jetton_wallets
-            .filter(jw => jw.balance && jw.balance !== '0')
-            .map(jw => {
-              const jettonRawAddress = jw.jetton
-              const jettonUserFriendly =
-                addressBook[jettonRawAddress]?.user_friendly ?? jettonRawAddress
-              const jettonMeta = metadata[jettonRawAddress]?.token_info?.[0]
-              const precision = jettonMeta?.extra?.decimals
-                ? parseInt(jettonMeta.extra.decimals, 10)
-                : 9
-
-              const assetId = toAssetId({
-                chainId: this.chainId,
-                assetNamespace: 'jetton',
-                assetReference: jettonUserFriendly,
-              })
-
-              return {
-                assetId,
-                balance: jw.balance,
-                symbol: jettonMeta?.symbol ?? '',
-                name: jettonMeta?.name ?? '',
-                precision,
-              }
-            })
-        }
+        tokens = buildTonTokens(jettonsResponse, this.chainId)
       } catch (err) {
         console.error('[TON] Error fetching jetton balances:', err)
         tokens = []
@@ -817,7 +411,7 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
               continue
             }
 
-            txsByTrace[traceId] = this.ownTraceTxs(trace, pubkey)
+            txsByTrace[traceId] = ownTraceTxs(trace, pubkey)
           }
         } catch (error) {
           console.error('[TON] Failed to resolve traces, dropping affected rows this page', {
@@ -853,13 +447,7 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
       for (const [traceId, traceGroup] of Object.entries(txsByTrace)) {
         const owner = traceGroup.find(t => t.hash === traceId) ?? traceGroup[0]
 
-        // Externally-initiated txs are keyed by their message hash - the same id broadcast
-        // returns and parseTx uses, so rows upserted at swap time overwrite history rows and
-        // vice versa instead of duplicating
-        const isExternalInitiated = !owner.in_msg?.source && Boolean(owner.in_msg?.hash)
-        const txid = base64ToHex(
-          isExternalInitiated && owner.in_msg?.hash ? owner.in_msg.hash : owner.hash,
-        )
+        const txid = getTraceOwnerTxid(owner)
 
         const allTransfers = buildTraceTransfers({
           txs: traceGroup,
@@ -915,10 +503,7 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
 
   async getSeqno(address: string): Promise<number> {
     try {
-      const result = await this.rpcRequest<{
-        exit_code: number
-        stack: [string, string][]
-      }>('runGetMethod', {
+      const result = await this.rpcRequest<TonRunGetMethodResult>('runGetMethod', {
         address,
         method: 'seqno',
         stack: [],
@@ -939,10 +524,7 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
   }
 
   async getJettonWalletAddress(jettonMaster: string, ownerAddress: string): Promise<string> {
-    const response = await this.httpApiRequest<{
-      jetton_wallets?: { address: string }[]
-      address_book?: Record<string, { user_friendly: string }>
-    }>(
+    const response = await this.httpApiRequest<TonJettonWalletsResponse>(
       `/api/v3/jetton/wallets?owner_address=${encodeURIComponent(
         ownerAddress,
       )}&jetton_address=${encodeURIComponent(jettonMaster)}&limit=1`,
@@ -1066,7 +648,7 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
     try {
       const { hex: signedTx } = input
 
-      const result = await this.rpcRequest<{ hash: string }>('sendBocReturnHash', {
+      const result = await this.rpcRequest<TonSendBocResult>('sendBocReturnHash', {
         boc: signedTx,
       })
 
@@ -1090,11 +672,9 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
       let storageFee = '0'
 
       try {
-        const configResult = await this.rpcRequest<{
-          gas_price?: string
-          flat_gas_limit?: string
-          flat_gas_price?: string
-        }>('getConfigParam', { config_id: 20 })
+        const configResult = await this.rpcRequest<TonConfigParamResult>('getConfigParam', {
+          config_id: 20,
+        })
 
         if (configResult.gas_price) {
           const gasPrice = BigInt(configResult.gas_price)
@@ -1158,12 +738,9 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
     try {
       const apiHash = isHexHash(msgHash) ? hexToBase64(msgHash) : msgHash
 
-      const result = await this.httpApiRequest<{
-        messages?: {
-          hash: string
-          in_msg_tx_hash?: string
-        }[]
-      }>(`/api/v3/messages?hash=${encodeURIComponent(apiHash)}`)
+      const result = await this.httpApiRequest<TonMessagesResponse>(
+        `/api/v3/messages?hash=${encodeURIComponent(apiHash)}`,
+      )
 
       if (!result.messages || result.messages.length === 0) {
         return TxStatus.Pending
@@ -1209,25 +786,13 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
     }
   }
 
-  private ownTraceTxs(trace: TonTrace, pubkey: string): TonTx[] {
-    return Object.values(trace.transactions ?? {})
-      .filter(t => addressesMatch(t.account, pubkey))
-      .sort((a, b) => (BigInt(a.lt) < BigInt(b.lt) ? -1 : 1))
-  }
-
   private async fetchJettonTransfers(
     pubkey: string,
     startLt: string,
     endLt: string,
-  ): Promise<{
-    jetton_transfers: JettonTransferRecord[]
-    address_book: Record<string, { user_friendly: string }>
-  }> {
+  ): Promise<Required<TonJettonTransfersResponse>> {
     try {
-      const response = await this.httpApiRequest<{
-        jetton_transfers?: JettonTransferRecord[]
-        address_book?: Record<string, { user_friendly: string }>
-      }>(
+      const response = await this.httpApiRequest<TonJettonTransfersResponse>(
         `/api/v3/jetton/transfers?owner_address=${encodeURIComponent(
           pubkey,
         )}&start_lt=${startLt}&end_lt=${endLt}&limit=100&sort=asc`,
@@ -1246,7 +811,6 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
     if (typeof txHashOrTx !== 'string') {
       throw new Error(`[TON] parseTx expects a string tx hash, got ${typeof txHashOrTx}`)
     }
-
     // status poll, history upsert and balance pipeline all parse the same hash within seconds -
     // a short-lived memo collapses them into one network run
     const cacheKey = `${txHashOrTx}:${pubkey}`
@@ -1295,9 +859,9 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
 
       if (!tx) {
         // Distinguish a known-but-unprocessed message (pending) from an unknown one
-        const msgResult = await this.httpApiRequest<{
-          messages?: { hash: string; in_msg_tx_hash?: string }[]
-        }>(`/api/v3/messages?hash=${encodeURIComponent(apiHash)}`)
+        const msgResult = await this.httpApiRequest<TonMessagesResponse>(
+          `/api/v3/messages?hash=${encodeURIComponent(apiHash)}`,
+        )
 
         if (!msgResult.messages || msgResult.messages.length === 0) {
           throw new Error('Message not found')
@@ -1333,7 +897,7 @@ export class ChainAdapter implements IChainAdapter<KnownChainIds.TonMainnet> {
       }
 
       const trace = traceResult.traces?.[0]
-      const traceTxs = trace ? this.ownTraceTxs(trace, pubkey) : []
+      const traceTxs = trace ? ownTraceTxs(trace, pubkey) : []
       const primaryTx = traceTxs[0] ?? tx
 
       const txsToProcess = traceTxs.length > 0 ? traceTxs : [tx]
